MINIFICPP-1443 Fix 'possible loss of data' warnings
... and a few others which were related or nearby in the code.


This closes #964

+++ controller/Controller.h
@@ -22,6 +22,7 @@
 #include "core/ConfigurationFactory.h"
 #include "io/ClientSocket.h"
 #include "c2/ControllerSocketProtocol.h"
+#include "utils/gsl.h"
 
 /**
  * Sends a single argument comment
@@ -35,7 +36,7 @@ bool sendSingleCommand(std::unique_ptr<minifi::io::Socket> socket, uint8_t op, c
   minifi::io::BufferStream stream;
   stream.write(&op, 1);
   stream.write(value);
-  return socket->write(const_cast<uint8_t*>(stream.getBuffer()), stream.size()) == stream.size();
+  return socket->write(const_cast<uint8_t*>(stream.getBuffer()), gsl::narrow<int>(stream.size())) == stream.size();
 }
 
 /**
@@ -76,7 +77,7 @@ int updateFlow(std::unique_ptr<minifi::io::Socket> socket, std::ostream &out, st
   stream.write(&op, 1);
   stream.write("flow");
   stream.write(file);
-  if (socket->write(const_cast<uint8_t*>(stream.getBuffer()), stream.size()) < 0) {
+  if (socket->write(const_cast<uint8_t*>(stream.getBuffer()), gsl::narrow<int>(stream.size())) < 0) {
     return -1;
   }
   // read the response
@@ -106,7 +107,7 @@ int getFullConnections(std::unique_ptr<minifi::io::Socket> socket, std::ostream 
   minifi::io::BufferStream stream;
   stream.write(&op, 1);
   stream.write("getfull");
-  if (socket->write(const_cast<uint8_t*>(stream.getBuffer()), stream.size()) < 0) {
+  if (socket->write(const_cast<uint8_t*>(stream.getBuffer()), gsl::narrow<int>(stream.size())) < 0) {
     return -1;
   }
   // read the response
@@ -132,7 +133,7 @@ int getJstacks(std::unique_ptr<minifi::io::Socket> socket, std::ostream &out) {
   minifi::io::BufferStream stream;
   stream.write(&op, 1);
   stream.write("jstack");
-  if (socket->write(const_cast<uint8_t*>(stream.getBuffer()), stream.size()) < 0) {
+  if (socket->write(const_cast<uint8_t*>(stream.getBuffer()), gsl::narrow<int>(stream.size())) < 0) {
     return -1;
   }
   // read the response
@@ -172,7 +173,7 @@ int getConnectionSize(std::unique_ptr<minifi::io::Socket> socket, std::ostream &
   stream.write(&op, 1);
   stream.write("queue");
   stream.write(connection);
-  if (socket->write(const_cast<uint8_t*>(stream.getBuffer()), stream.size()) < 0) {
+  if (socket->write(const_cast<uint8_t*>(stream.getBuffer()), gsl::narrow<int>(stream.size())) < 0) {
     return -1;
   }
   // read the response
@@ -192,7 +193,7 @@ int listComponents(std::unique_ptr<minifi::io::Socket> socket, std::ostream &out
   uint8_t op = minifi::c2::Operation::DESCRIBE;
   stream.write(&op, 1);
   stream.write("components");
-  if (socket->write(const_cast<uint8_t*>(stream.getBuffer()), stream.size()) < 0) {
+  if (socket->write(const_cast<uint8_t*>(stream.getBuffer()), gsl::narrow<int>(stream.size())) < 0) {
     return -1;
   }
   uint16_t responses = 0;
@@ -216,7 +217,7 @@ int listConnections(std::unique_ptr<minifi::io::Socket> socket, std::ostream &ou
   uint8_t op = minifi::c2::Operation::DESCRIBE;
   stream.write(&op, 1);
   stream.write("connections");
-  if (socket->write(const_cast<uint8_t*>(stream.getBuffer()), stream.size()) < 0) {
+  if (socket->write(const_cast<uint8_t*>(stream.getBuffer()), gsl::narrow<int>(stream.size())) < 0) {
     return -1;
   }
   uint16_t responses = 0;
